Refactor dashboard and related components for improved modularity
- Introduce ActionButtons component to handle both share and download actions.

- Remove obsolete components (CryptoDetails, CryptoTable) and replace with SelectedCryptoDetails.

- Delete unused hook (useCryptoData) and streamline dashboard.tsx logic.

- Update globals.css and tsconfig.json settings to align with new code structure.

- Convert JavaScript utils to TypeScript (firebase-config, requireAuth) for type safety.

- Add new utility files (formatting.ts, sorting.ts, types.ts) to segregate logic and types.

These changes enhance readability, maintainability, and scalability of the codebase while aligning with best practices in modularity and type safety.
