+++ answers/amal_mohammed/utils/types.ts
@@ -1,0 +1,9 @@
+// utils/types.ts
+export interface CryptoDetail {
+    s: string;  // Symbol
+    c: string;  // Last price
+    P: string;  // 24-hour change percentage
+    h: string;  // 24-hour high
+    l: string;  // 24-hour low
+    v: string;  // 24-hour volume
+}
